+++ src/main/resources/templates/index.html
@@ -348,7 +348,7 @@
 </div>
 
 <a href="test">클릭</a>&nbsp; &nbsp; &nbsp; &nbsp;
-<a href="/psychological/test">클릭</a>&nbsp; &nbsp; &nbsp; &nbsp;
+<a href="/psychometry/test">클릭</a>&nbsp; &nbsp; &nbsp; &nbsp;
 <a href="/review/reviewmain">후기</a>&nbsp; &nbsp; &nbsp; &nbsp;
 <a href="/admin/editUserInfo">관리자용/회원수정</a>&nbsp; &nbsp; &nbsp; &nbsp;
 <a href="/admin/editTherapistInfo">관리자용/상담사수정</a>&nbsp; &nbsp; &nbsp; &nbsp;
