+++ src/main/resources/templates/psychological/psychometry/firstTest.html
@@ -36,7 +36,7 @@
     </section>
     <br>
     <!-- 검사하기 버튼 -->
-    <button class="roundbutton" onclick="location.href='/psychological/start'">검사하기</button>
+    <button class="roundbutton" onclick="location.href='/psychological/psychometry/firstTest'">검사하기</button>
     <br><br><br>
 </div>
 <!-- 풋터 -->
