+++ src/main/resources/templates/psychological/psychometry/lastTest.html
@@ -14,7 +14,7 @@
 <div class="box"><br>
     해당사항이 있으면 체크해주세요.<br><br>
     <h3>희망 상담사 성별</h3>
-    <form action="/psychological/loding" method="get"><br>
+    <form action="/psychometry/loding" method="get"><br>
         <input type="radio" name="gender" value="남">남자<br>
         <input type="radio" name="gender" value="여">여자<br>
         <input type="radio" name="gender" value="무관">무관<br><br>
