+++ src/main/resources/templates/psychological/psychometry/resultTest.html
@@ -15,7 +15,7 @@
 
     <div class="wrap">
         <div class="wrap img">
-            <img class="grass" src="/images/common/Background Shape.png" alt="grass">
+            <img class="grass" src="/images/common/Background%20Shape.png" alt="grass">
         </div>
             <div class="wraptext">
             <div>
